docs: document implementation matrix
The idea w/ this matrix is to provide end users a clear picture of what is implemented
for each language implementation. Ideally, the matrix will be:

1. auto generated from the kitchen-sink
2. automatically filled in for every implementation via a test

In the short term, the automation for fill in the implementation details could be manual, but
ideally that would be short-lived for the sake of users and implementors.

Signed-off-by: Alex Boten <223565+codeboten@users.noreply.github.com>